Fixed obscure ForEach looping bug with expanding array
Elements substitution was incomplete when multiple elements were involved : If multiple substitutions of the same element occured when the format had been already splitted into multiple nodes, the last substitution(s) wouldn’t be applied as the last segment(s) wouldn’t be inspected.

ForEach evaluated the length of the array at once and the array was expanded during the substitutions.

The order of the elements in the new test is important: first [br] splits the format node into a 3-node array then [strong] should sustitute nodes to the first and last nodes.

Renamed a few variables as well to enhance readability.

Signed-off-by: Frédéric Miserey <frederic@none.net>